move CellStatus and adcVoltages to COMMON::types
