+++ libs/balance_port/src/balance_port.cpp
@@ -18,8 +18,8 @@ bool BalancePort::initADC(i2c_inst_t* i2c_port) {
     }
 }
 
-adcVoltages BalancePort::getCellVoltages() {
-    adcVoltages voltages;
+COMMON::adcVoltages BalancePort::getCellVoltages() {
+    COMMON::adcVoltages voltages;
     int16_t adc0, adc1, adc2, adc3;
 
     adc0 = inputVoltagesADC.readADC_SingleEnded(ADSX_AIN0);
@@ -35,8 +35,8 @@ adcVoltages BalancePort::getCellVoltages() {
     return voltages;
 }
 
-CellStatus BalancePort::checkVoltages(adcVoltages measuredVoltages) {
-    CellStatus voltageStatus;
+COMMON::CellStatus BalancePort::checkVoltages(COMMON::adcVoltages measuredVoltages) {
+    COMMON::CellStatus voltageStatus;
     voltageStatus.voltages =  measuredVoltages;
     float cell1 = measuredVoltages.cell1;
     float cell2 = measuredVoltages.cell2;
@@ -49,29 +49,29 @@ CellStatus BalancePort::checkVoltages(adcVoltages measuredVoltages) {
     if ((maxVoltage - minVoltage) > balanceThreshold){
         voltageStatus.outOfBalance = true;
         voltageStatus.allOk = false;
-        voltageStatus.fault = "cells out of balance, ";
+        // voltageStatus.fault = "cells out of balance, ";
     }
     if ( minVoltage < minCellVoltage ){
         voltageStatus.lowCellVoltage = true;
         voltageStatus.allOk = false;
-        voltageStatus.fault += "cell undervoltage, ";
+        // voltageStatus.fault += "cell undervoltage, ";
     }
     if (maxVoltage > maxCellVoltage) {
         voltageStatus.highCellVoltage = true;
         voltageStatus.allOk = false;
-        voltageStatus.fault += "cell overvoltage, ";
+        // voltageStatus.fault += "cell overvoltage, ";
     }
     if (PSU > PSUConnectedThreshold && PSU < minPSU) {
         voltageStatus.psuUnderVoltage = true;
         voltageStatus.allOk = false;
-        voltageStatus.fault += "PSU undervoltage, ";
+        // voltageStatus.fault += "PSU undervoltage, ";
     }
     return voltageStatus;
 }
 
 void BalancePort::raiseCellStatus() {
-    const adcVoltages voltages = getCellVoltages(); // Assume this method exists and fetches voltages
-    const CellStatus status = checkVoltages(voltages);
+    const COMMON::adcVoltages voltages = getCellVoltages(); // Assume this method exists and fetches voltages
+    const COMMON::CellStatus status = checkVoltages(voltages);
     if (!status.allOk) {
         // Increment failCount if the cell status is not okay for 10 consecutive times
         failCount++;
